# crates/eidolons-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eidolons API",
    "description": "Privacy-transparent AI proxy API with inline verification metadata",
    "contact": {
      "name": "Mike Marcacci"
    },
    "license": {
      "name": "",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/account/tokens": {
      "post": {
        "tags": [
          "Linked"
        ],
        "summary": "`POST /v1/account/tokens` — issue anonymous credit tokens (authenticated).",
        "operationId": "issue_tokens",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueTokensRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueTokensResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credit balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Token issuance not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "message",
          "type"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "An error response in OpenAI format.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "IssueTokensRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/account/tokens`.",
        "required": [
          "issuance_request",
          "credits"
        ],
        "properties": {
          "credits": {
            "type": "integer",
            "format": "int64",
            "description": "Number of credits to issue."
          },
          "issuance_request": {
            "type": "string",
            "description": "Base64-encoded CBOR `IssuanceRequest`."
          }
        }
      },
      "IssueTokensResponse": {
        "type": "object",
        "description": "Response for `POST /v1/account/tokens`.",
        "required": [
          "issuance_response",
          "epoch",
          "credits",
          "ledger_entry_id"
        ],
        "properties": {
          "credits": {
            "type": "integer",
            "format": "int64",
            "description": "Number of credits issued."
          },
          "epoch": {
            "type": "string",
            "description": "Epoch identifier (YYYY-MM)."
          },
          "issuance_response": {
            "type": "string",
            "description": "Base64-encoded CBOR `IssuanceResponse`."
          },
          "ledger_entry_id": {
            "type": "string",
            "description": "The ledger entry ID for this issuance."
          }
        }
      },
...
    },
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic",
        "description": "Account credentials. Username is the account UUID, password is the secret returned at account creation."
      }
    }
  },
...
}